Fixed version manager constructor.

+++ .program-files/modules/version.py
@@ -3,7 +3,7 @@ from time import time
 
 
 class VersionManager:
-    def __init__(self, daysInCache: int) -> None:
+    def __init__(self, daysInCache: float) -> None:
         self.versionPath = path.abspath(
             path.join(path.dirname(__file__), '../config/VERSION'))
         self.daysInCache = daysInCache if daysInCache > 0 else 1
